refactor: update estimation functions for token and NFT transfers
