+++ packages/core-js/src/userop/index.ts
@@ -4,9 +4,9 @@ export { tokenTransfer } from './tokenTransfer.js';
 export { nftCreateCollection } from './nftCollectionCreation.js';
 export { nftMint } from './nftMint.js';
 export { aptosRawTransaction } from './aptosRawTransaction.js';
-export { estimateTokenTransfer } from './estimateTokenTransfer.js';
-export { estimateNftTransfer } from './estimateNftTransfer.js';
-export { estimateEvmRawTransaction } from './estimateEvmRawTransaction.js';
-export { estimateAptosRawTransaction } from './estimateAptosRawTransaction.js';
-export { estimateNftMint } from './estimateNftMint.js';
-export { estimateNftCreateCollection } from './estimateNftCollectionCreation.js';
+export { tokenTransferWithEstimate } from './tokenTransferWithEstimate.js';
+export { nftTransferWithEstimate } from './nftTransferWithEstimate.js';
+export { evmRawTransactionWithEstimate } from './evmRawTransactionWithEstimate.js';
+export { aptosRawTransactionWithEstimate } from './aptosRawTransactionWithEstimate.js';
+export { nftMintWithEstimate } from './nftMintWithEstimate.js';
+export { nftCreateCollectionWithEstimate } from './nftCreateCollectionWithEstimate.js';
