+++ packages/core-js/src/userop/nftCreateCollectionWithEstimate.ts
@@ -12,61 +12,65 @@ import {
   toHex,
 } from 'viem';
 import { INTENT_ABI } from './abi.js';
-import BffClientRepository from '@/api/bff.js';
+import type { NftCreateCollectionParams } from './types.js';
 import {
-  TokenTransferIntentParamsSchema,
+  NftCreateCollectionParamsSchema,
   validateSchema,
 } from './userOpInputValidator.js';
-import type { TokenTransferIntentParams } from './types.js';
+import BffClientRepository from '@/api/bff.js';
 import type {
   EstimationDetails,
-  TokenTransferEstimateRequest,
+  NftCreateCollectionEstimateRequest,
 } from '@/types/bff/estimate.js';
 
 /**
- * Creates a user operation for token transfer.
+ * Creates a user operation for NFT collection creation.
  *
- * This function initiates the process of transferring tokens by encoding
+ * This function initiates the process of creating an NFT collection by encoding
  * the necessary parameters into a User Operation. The operation is then
  * submitted through the OktoClient for execution.
  *
  * @param oc - The OktoClient instance used to interact with the blockchain.
- * @param data - The parameters for transferring tokens.
- * @param feePayerAddress - Optional fee payer address, defaults to Constants.FEE_PAYER_ADDRESS.
- * @returns The User Operation (UserOp) for the token transfer and transfer details.
+ * @param data - The parameters for creating the NFT collection (caip2Id, name, uri, and optional data with attributes, symbol, type, description).
+ * @param feePayerAddress - Optional fee payer address.
+ * @returns The User Operation (UserOp) for the NFT collection creation and collection details.
  */
-export async function estimateTokenTransfer(
+export async function nftCreateCollectionWithEstimate(
   oc: OktoClient,
-  data: TokenTransferIntentParams,
+  data: NftCreateCollectionParams,
   feePayerAddress?: Address,
 ): Promise<{ userOp: UserOp; details: EstimationDetails }> {
   if (!oc.isLoggedIn()) {
     throw new BaseError('User not logged in');
   }
+  validateSchema(NftCreateCollectionParamsSchema, data);
 
-  validateSchema(TokenTransferIntentParamsSchema, data);
+  const nonce = generateUUID();
 
   if (!feePayerAddress) {
     feePayerAddress = Constants.FEE_PAYER_ADDRESS;
   }
 
-  if (data.recipient === oc.userSWA) {
-    throw new BaseError('Recipient address cannot be same as the user address');
-  }
-
-  const nonce = generateUUID();
-
   const gasPrice = await GatewayClientRepository.getUserOperationGasPrice(oc);
 
   const chains = await getChains(oc);
-  const chain = chains.find(
+  const currentChain = chains.find(
     (chain) => chain.caipId.toLowerCase() === data.caip2Id.toLowerCase(),
   );
 
-  if (!chain) {
+  if (!currentChain) {
     throw new BaseError(`Chain Not Supported`, {
       details: `${data.caip2Id} is not supported for this client`,
     });
+  }
+
+  if (!currentChain.caipId.toLowerCase().startsWith('aptos:')) {
+    throw new BaseError(
+      'NFT Collection creation is only supported on Aptos chain',
+      {
+        details: `Provided chain: ${currentChain.caipId}`,
+      },
+    );
   }
 
   const paymasterData = await oc.paymasterData({
@@ -74,34 +78,40 @@ export async function estimateTokenTransfer(
     validUntil: new Date(Date.now() + 6 * Constants.HOURS_IN_MS),
   });
 
-  const requestBody: TokenTransferEstimateRequest = {
-    type: Constants.INTENT_TYPE.TOKEN_TRANSFER,
+  // Prepare and send estimation request
+  const requestBody: NftCreateCollectionEstimateRequest = {
+    type: Constants.INTENT_TYPE.NFT_CREATE_COLLECTION,
     jobId: nonce,
     paymasterData,
     gasDetails: {
       maxFeePerGas: gasPrice.maxFeePerGas,
       maxPriorityFeePerGas: gasPrice.maxPriorityFeePerGas,
     },
-    details: {
-      recipientWalletAddress: data.recipient,
-      caip2Id: data.caip2Id,
-      tokenAddress: data.token,
-      amount: data.amount.toString(),
-    },
+    feePayerAddress,
+    details: data,
   };
 
   // Get estimate from BFF API
-  const transferEstimate = await BffClientRepository.getTokenTransferEstimate(
+  const nftEstimate = await BffClientRepository.getNftCreateCollectionEstimate(
     oc,
     requestBody,
   );
 
+  const nftData = JSON.stringify({
+    type: data.data.type || '',
+    attributes: data.data.attributes || '',
+    description: data.data.description || '',
+    symbol: data.data.symbol || '',
+  });
+
+  const nftDataEncoded = toHex(new TextEncoder().encode(nftData));
+
   const jobParametersAbiType =
-    '(string caip2Id, string recipientWalletAddress, string tokenAddress, uint amount)';
+    '(string caip2Id, string name, string uri, bytes data)';
   const gsnDataAbiType = `(bool isRequired, string[] requiredNetworks, ${jobParametersAbiType}[] tokens)`;
 
   const calldata = encodeAbiParameters(
-    parseAbiParameters('bytes4, address,uint256, bytes'),
+    parseAbiParameters('bytes4, address, uint256, bytes'),
     [
       Constants.EXECUTE_USEROP_FUNCTION_SELECTOR,
       oc.env.jobManagerAddress,
@@ -118,8 +128,8 @@ export async function estimateTokenTransfer(
             parseAbiParameters('(bool gsnEnabled, bool sponsorshipEnabled)'),
             [
               {
-                gsnEnabled: chain.gsnEnabled ?? false,
-                sponsorshipEnabled: chain.sponsorshipEnabled ?? false,
+                gsnEnabled: currentChain.gsnEnabled ?? false,
+                sponsorshipEnabled: currentChain.sponsorshipEnabled ?? false,
               },
             ],
           ),
@@ -132,13 +142,13 @@ export async function estimateTokenTransfer(
           ]),
           encodeAbiParameters(parseAbiParameters(jobParametersAbiType), [
             {
-              amount: BigInt(data.amount),
               caip2Id: data.caip2Id,
-              recipientWalletAddress: data.recipient,
-              tokenAddress: data.token,
+              name: data.name,
+              uri: data.uri,
+              data: nftDataEncoded,
             },
           ]),
-          Constants.INTENT_TYPE.TOKEN_TRANSFER,
+          Constants.INTENT_TYPE.NFT_CREATE_COLLECTION,
         ],
       }),
     ],
@@ -149,25 +159,25 @@ export async function estimateTokenTransfer(
     nonce: toHex(nonceToBigInt(nonce), { size: 32 }),
     paymaster: oc.env.paymasterAddress,
     callGasLimit:
-      transferEstimate.userOps.callGasLimit ||
+      nftEstimate.userOps.callGasLimit ||
       toHex(Constants.GAS_LIMITS.CALL_GAS_LIMIT),
     verificationGasLimit:
-      transferEstimate.userOps.verificationGasLimit ||
+      nftEstimate.userOps.verificationGasLimit ||
       toHex(Constants.GAS_LIMITS.VERIFICATION_GAS_LIMIT),
     preVerificationGas:
-      transferEstimate.userOps.preVerificationGas ||
+      nftEstimate.userOps.preVerificationGas ||
       toHex(Constants.GAS_LIMITS.PRE_VERIFICATION_GAS),
     maxFeePerGas: gasPrice.maxFeePerGas,
     maxPriorityFeePerGas: gasPrice.maxPriorityFeePerGas,
     paymasterPostOpGasLimit:
-      transferEstimate.userOps.paymasterPostOpGasLimit ||
+      nftEstimate.userOps.paymasterPostOpGasLimit ||
       toHex(Constants.GAS_LIMITS.PAYMASTER_POST_OP_GAS_LIMIT),
     paymasterVerificationGasLimit:
-      transferEstimate.userOps.paymasterVerificationGasLimit ||
+      nftEstimate.userOps.paymasterVerificationGasLimit ||
       toHex(Constants.GAS_LIMITS.PAYMASTER_VERIFICATION_GAS_LIMIT),
-    callData: transferEstimate.userOps.callData || calldata,
+    callData: nftEstimate.userOps.callData || calldata,
     paymasterData:
-      transferEstimate.userOps.paymasterData ||
+      nftEstimate.userOps.paymasterData ||
       (await oc.paymasterData({
         nonce: nonce,
         validUntil: new Date(Date.now() + 6 * Constants.HOURS_IN_MS),
@@ -176,6 +186,6 @@ export async function estimateTokenTransfer(
 
   return {
     userOp,
-    details: transferEstimate.details,
+    details: nftEstimate.details,
   };
 }
