+++ packages/react-native/src/userop/index.ts
@@ -5,12 +5,12 @@ import {
   nftCreateCollection,
   nftMint,
   aptosRawTransaction,
-  estimateTokenTransfer,
-  estimateNftTransfer,
-  estimateEvmRawTransaction,
-  estimateAptosRawTransaction,
-  estimateNftCreateCollection,
-  estimateNftMint,
+  tokenTransferWithEstimate,
+  nftTransferWithEstimate,
+  evmRawTransactionWithEstimate,
+  aptosRawTransactionWithEstimate,
+  nftMintWithEstimate,
+  nftCreateCollectionWithEstimate,
 } from '@okto_web3/core-js-sdk/userop';
 
 export {
@@ -20,10 +20,10 @@ export {
   nftCreateCollection,
   nftMint,
   aptosRawTransaction,
-  estimateTokenTransfer,
-  estimateNftTransfer,
-  estimateEvmRawTransaction,
-  estimateAptosRawTransaction,
-  estimateNftCreateCollection,
-  estimateNftMint,
+  tokenTransferWithEstimate,
+  nftTransferWithEstimate,
+  evmRawTransactionWithEstimate,
+  aptosRawTransactionWithEstimate,
+  nftMintWithEstimate,
+  nftCreateCollectionWithEstimate,
 };
